requirements.txt
ccxt
pandas
numpy
scikit-learn
ta  # technical analysis helpers
matplotlib
python-dotenv
joblib
